[#19] feat: seperate PropListView into PropsListView and PropsGalleryView and rename PropListView to ScenarioPropsVIew
